Started Flatland 2D
Begun work on the game Flatland 2D. A tile based 2D RPG.
Made a simple square sprite image to be the hero and
a quick pygame while loop to display him. Also added a
txt file of ideas for this game so I can keep track of
the good/bad things as I go along. Beware Spoilers in that
file.