Organisation changes
Have added lots of config files for setup and sharing common
objects. In doing so I've changed the size of the main screen
and tile size to be slightly larger than the hero sprite.
To prevent him being the same size as the tiles.
Also added an mpty sprite base class file for sprite movement and
behaviour.