Config changes and player movement
Have moved the player animation to its own class. So now
the sprit changes drection based on the current
movement keypress.
Movement commands are in the main config file.
Certain tile config values have been moved around, and a
more sensible tilesize compared to hero sprite size eqution
is used.
Number of tiles is now odd, to allow the hero to constantly be in the
middle of the screen.