Huge update:
Changed a whole bunch of stuff to do with the movement
of the player. Now has a basic animation while moving,
though the maths of this will need tweaking.
Added gamestates for pausing, moving, rotation etc. These
still need fiddling with to prevent movement events from stacking
while moving.
Added a utils.py file to do tuple maths including some coordinate
transformations.
There was a huge bug for a while with moving downwards, but it's fixed
now ('is not' conditional was wrongly used). Now there's still a bug with
the tile placement on the map/tile index. The player sees tiles in the wrong
places.