Rotation and Movement fixes + small map update
Added water around the outer edge of the level so we can't get to
the out of index areas. Required a quick change to the start position.

Made a velocity variable for the player so the moving speed can be
constant. Need to make a speed config variable (~0.1)

Now check for closeness of current angle to target angle and
just set the to equal to oine another if close. Preventing any
appearence of 'wiggling' towards the target angle. Now need
to correctly rotate the player and we can then merge back to
the master branch.