Level Image
**This commit doesn't work**
Have started the move to making the background layer of
the level into one large image. Hopefully scrolling + blitting
this will be faster than doing so for the individual tiles
every frame. The image processing is currently done, so the
Map() class makes the correct image but the game hasn't
been changed to take advantage of it yet.