Level Working
Have changed the Map and GameScreen classes to use the new
large image for blitting. Seems faster but still quite a bit of
slowdown when using larger levels. Investigate scrolling rather
than changing the position fully. This has also gotten rid
of the black lines due to pixel rounding errors, so we'll
keep it.
Have also added an fps measurement to the screen and have
made the game use only one instance of Clock() as I think
pygame didn't like that I was using several for different
gamestates.